+++ src/components/feedback/__tests__/feedback-sent-rating.test.js
@@ -6,7 +6,99 @@ import Adapter from 'enzyme-adapter-react-16';
 
 configure({ adapter: new Adapter() });
 
-describe('Sent helpful rating', () => {
+describe('Sent helpful rating - yes', () => {
+  const feedback = mount(
+    <Feedback
+      site="dr-ui"
+      location={{
+        pathname: '/mapbox-gl-js/api/',
+        hash: '#lnglat'
+      }}
+      userName="decorah"
+      webhook={{
+        production: '',
+        staging: ''
+      }}
+    />
+  );
+  test('clicked yes', () => {
+    feedback.find('#dr-ui--feedback-page-yes').simulate('click');
+    expect(feedback.state()).toEqual({
+      event: {
+        event: 'Sent docs feedback',
+        userId: 'decorah',
+        properties: {
+          environment: 'staging',
+          helpful: true,
+          location: {
+            hash: '',
+            host: 'localhost',
+            hostname: 'localhost',
+            href: 'http://localhost/',
+            origin: 'http://localhost',
+            pathname: '/',
+            search: ''
+          },
+          page: { hash: '#lnglat', pathname: '/mapbox-gl-js/api/' },
+          section: undefined,
+          site: 'dr-ui',
+          userId: 'decorah'
+        }
+      },
+      feedback: undefined,
+      feedbackSent: undefined,
+      helpful: true
+    });
+  });
+});
+
+describe('Sent helpful rating - no', () => {
+  const feedback = mount(
+    <Feedback
+      site="dr-ui"
+      location={{
+        pathname: '/mapbox-gl-js/api/',
+        hash: '#lnglat'
+      }}
+      userName="decorah"
+      webhook={{
+        production: '',
+        staging: ''
+      }}
+    />
+  );
+  test('clicked no', () => {
+    feedback.find('#dr-ui--feedback-page-no').simulate('click');
+    expect(feedback.state()).toEqual({
+      event: {
+        event: 'Sent docs feedback',
+        userId: 'decorah',
+        properties: {
+          environment: 'staging',
+          helpful: false,
+          location: {
+            hash: '',
+            host: 'localhost',
+            hostname: 'localhost',
+            href: 'http://localhost/',
+            origin: 'http://localhost',
+            pathname: '/',
+            search: ''
+          },
+          page: { hash: '#lnglat', pathname: '/mapbox-gl-js/api/' },
+          section: undefined,
+          site: 'dr-ui',
+          userId: 'decorah'
+        }
+      },
+      feedback: undefined,
+      feedbackSent: undefined,
+      helpful: false
+    });
+  });
+});
+
+describe('Sent helpful rating snapshot', () => {
   const feedback = shallow(
     <Feedback
       site="dr-ui"
@@ -20,7 +112,7 @@ describe('Sent helpful rating', () => {
     helpful: true
   });
 
-  test('basic', () => {
+  test('snapshot', () => {
     expect(toJson(feedback)).toMatchSnapshot();
   });
 });
@@ -38,17 +130,19 @@ describe('Feedback character limit', () => {
     });
     // make sure submit button is not disabled
     const submitBtn = feedback
-      .find('#dr-ui--feedback-submit-button')
+      .find('#dr-ui--feedback-page-submit')
       .at(0)
       .props();
     expect(submitBtn.disabled).toBe(false);
 
     // make sure overlimit warning is not shown
-    const overlimitWarning = feedback.find('#dr-ui--feedback-overlimit').at(0);
+    const overlimitWarning = feedback
+      .find('#dr-ui--feedback-page-overlimit')
+      .at(0);
     expect(overlimitWarning.exists()).toBe(false);
 
     // make sure character counter is equal
-    const charCounter = feedback.find('#dr-ui--feedback-char-counter').at(0);
+    const charCounter = feedback.find('#dr-ui--feedback-page-counter').at(0);
     expect(charCounter.props().children).toEqual(974);
     // snapshot
     expect(toJson(feedback)).toMatchSnapshot();
@@ -62,17 +156,19 @@ describe('Feedback character limit', () => {
     });
     // make sure submit button is disabled
     const submitBtn = feedback
-      .find('#dr-ui--feedback-submit-button')
+      .find('#dr-ui--feedback-page-submit')
       .at(0)
       .props();
     expect(submitBtn.disabled).toBe(true);
 
     // make sure overlimit warning is shown
-    const overlimitWarning = feedback.find('#dr-ui--feedback-overlimit').at(0);
+    const overlimitWarning = feedback
+      .find('#dr-ui--feedback-page-overlimit')
+      .at(0);
     expect(overlimitWarning.exists()).toBe(true);
 
     // make sure character counter is equal
-    const charCounter = feedback.find('#dr-ui--feedback-char-counter').at(0);
+    const charCounter = feedback.find('#dr-ui--feedback-page-counter').at(0);
     expect(charCounter.props().children).toEqual(-340);
     // snapshot
     expect(toJson(feedback)).toMatchSnapshot();