+++ src/components/feedback/__tests__/feedback-test-cases.js
@@ -85,7 +85,7 @@ testCases.common = {
       </TopbarSticker>
       <div className="limiter">
         <PageLayout
-          sidebarTitle={<div className="ml36">Section title</div>}
+          sidebarTitle="Section title"
           sidebarContent={
             <NavigationAccordion
               currentPath="page-one"