+++ src/components/sectioned-navigation/__tests__/sectioned-navigation-test-cases.js
@@ -1,4 +1,7 @@
+import React from 'react';
 import SectionedNavigation from '../sectioned-navigation';
+import PageLayout from '../../page-layout/page-layout';
+import examples from './examples';
 
 const testCases = {};
 
@@ -11,6 +14,7 @@ testCases.basic = {
       {
         title: 'Getting started',
         url: '#getting-started',
+        id: 'getting-started',
         items: [
           {
             text: 'Camera animation',
@@ -30,6 +34,7 @@ testCases.basic = {
       {
         title: 'Markers and callouts',
         url: '#markers',
+        id: 'markers',
         items: [
           {
             text: 'Annotation models',
@@ -44,6 +49,7 @@ testCases.basic = {
       {
         title: 'Getting started again',
         url: '#getting-started-again',
+        id: 'getting-started-again',
         items: [
           {
             text: 'Apply a style',
@@ -94,6 +100,7 @@ testCases.minimal = {
       {
         title: 'Getting started again',
         url: '#getting-started-again',
+        id: 'getting-started-again',
         items: [
           {
             text: 'Apply a style',
@@ -114,6 +121,7 @@ testCases.headingsOnly = {
       {
         title: 'Getting started',
         url: '#getting-started',
+        id: 'getting-started',
         items: [
           {
             text: 'Camera animation',
@@ -132,6 +140,7 @@ testCases.headingsOnly = {
       {
         title: 'Markers and callouts',
         url: '#markers-and-callouts',
+        id: 'markers-and-callouts',
         items: [
           {
             text: 'Annotation models',
@@ -146,6 +155,7 @@ testCases.headingsOnly = {
       {
         title: 'Getting started again',
         url: '#getting-started-again',
+        id: 'getting-started-again',
         items: [
           {
             text: 'Apply a style',
@@ -167,6 +177,7 @@ testCases.filter = {
       {
         title: 'Getting started',
         url: '#getting-started',
+        id: 'getting-started',
         items: [
           {
             text: 'Camera animation',
@@ -188,6 +199,7 @@ testCases.filter = {
       {
         title: 'Markers and callouts',
         url: '#markers',
+        id: 'markers',
         items: [
           {
             text: 'Annotation models',
@@ -204,6 +216,7 @@ testCases.filter = {
       {
         title: 'Getting started again',
         url: '#getting-started-again',
+        id: 'getting-started-again',
         items: [
           {
             text: 'Apply a style',
@@ -213,6 +226,23 @@ testCases.filter = {
         ]
       }
     ]
+  }
+};
+
+testCases.page = {
+  description: 'PageLayout',
+  component: PageLayout,
+  props: {
+    sidebarContent: (
+      <SectionedNavigation
+        title="Examples"
+        includeFilterBar={true}
+        sections={examples}
+      />
+    ),
+    sidebarContentStickyTop: 0,
+    sidebarContentStickyTopNarrow: 0,
+    children: <div>Doc content</div>
   }
 };